+++ src/Query/Expr/StrictlyLower.php
@@ -1,0 +1,54 @@
+<?php
+
+/*
+ * East Common.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license
+ * it is available in LICENSE file at the root of this package
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richard@teknoo.software so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ *
+ * @link        https://teknoo.software/east-collection/common Project website
+ *
+ * @license     https://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+
+declare(strict_types=1);
+
+namespace Teknoo\East\Common\Query\Expr;
+
+use Teknoo\East\Common\Contracts\Query\Expr\ExprInterface;
+use Teknoo\Immutable\ImmutableInterface;
+use Teknoo\Immutable\ImmutableTrait;
+
+/**
+ * Operator to add a constraint to check if a value is not equal to a value
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ * @license     https://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+class StrictlyLower implements ExprInterface, ImmutableInterface
+{
+    use ImmutableTrait;
+
+    public function __construct(
+        private readonly mixed $value,
+    ) {
+        $this->uniqueConstructorCheck();
+    }
+
+    public function getValue(): mixed
+    {
+        return $this->value;
+    }
+}