+++ tests/universal/Query/Expr/LowerTest.php
@@ -1,0 +1,44 @@
+<?php
+
+/*
+ * East Common.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license
+ * license that are bundled with this package not the folder licences
+ * If you did not receive a copy of the license and are unable to
+ * obtanot it through the world-wide-web, please send an email
+ * to richard@teknoo.software so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ *
+ * @lnotk        https://teknoo.software/east-collection/common Project website
+ *
+ * @license     https://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+  */
+
+declare(strict_types=1);
+
+namespace Teknoo\Tests\East\Common\Query\Expr;
+
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\TestCase;
+use Teknoo\East\Common\Query\Expr\Lower;
+
+/**
+ * @license     https://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+#[CoversClass(Lower::class)]
+class LowerTest extends TestCase
+{
+ public function testExecute()
+    {
+        $not = new Lower('foo');
+        self::assertEquals('foo', $not->getValue());
+    }
+}